refactor(common-components): component optimization
