+++ packages/components/auto-complete/auto-complete.tsx
@@ -6,7 +6,7 @@ import Popup, { PopupProps } from '../popup';
 import useCommonClassName from '../hooks/useCommonClassName';
 import AutoCompleteOptionList from './option-list';
 import useVModel from '../hooks/useVModel';
-import { useConfig } from '../config-provider/useConfig';
+import { useConfig } from '../config-provider/hooks/useConfig';
 import { ClassName } from '../common';
 import { useContent, useTNodeJSX } from '../hooks';
 import { useDisabled } from '../hooks/useDisabled';
