+++ packages/components/common-components/fake-arrow.tsx
@@ -26,21 +26,16 @@ export default defineComponent({
       },
       props.overlayClassName,
     ]);
-    return {
-      classes,
-    };
-  },
 
-  render() {
-    return (
+    return () => (
       <svg
-        class={this.classes}
+        class={classes.value}
         width="16"
         height="16"
         viewBox="0 0 16 16"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
-        style={this.overlayStyle}
+        style={props.overlayStyle}
       >
         <path d="M3.75 5.7998L7.99274 10.0425L12.2361 5.79921" stroke="black" stroke-opacity="0.9" stroke-width="1.3" />
       </svg>
