+++ packages/components/config-provider/config-provider.tsx
@@ -1,7 +1,7 @@
 import { defineComponent } from 'vue';
 import props from './props';
 import { useTNodeJSX } from '../hooks/tnode';
-import { provideConfig } from './useConfig';
+import { provideConfig } from './hooks/useConfig';
 
 export default defineComponent({
   name: 'TConfigProvider',
