+++ packages/components/config-provider/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useConfig';
