+++ packages/components/config-provider/utils/context.ts
@@ -1,10 +1,8 @@
 import { InjectionKey, ComputedRef } from 'vue';
-import { mergeWith as _mergeWith } from 'lodash-es';
-import { merge } from 'lodash-es';
+import { mergeWith as _mergeWith, merge, isArray } from 'lodash-es';
 import defaultConfig from '@tdesign/common-js/global-config/default-config';
 import defaultZhLocale from '@tdesign/common-js/global-config/locale/zh_CN';
-import { GlobalConfigProvider } from './type';
-import { isArray } from 'lodash-es';
+import { GlobalConfigProvider } from '../type';
 
 export enum EAnimationType {
   ripple = 'ripple',
