+++ packages/components/config-provider/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './context';
