+++ packages/components/typography/ellipsis.tsx
@@ -2,7 +2,7 @@ import { defineComponent, computed, ref } from 'vue';
 import { usePrefixClass } from '../hooks/useConfig';
 import props from './paragraph-props';
 import TTooltip from '../tooltip/index';
-import { useConfig } from '../config-provider/useConfig';
+import { useConfig } from '../config-provider/hooks/useConfig';
 
 import type { TypographyEllipsis } from './type';
